refactored pipeline building to builder
