+++ app/src/VulkanBackend.hpp
@@ -10,6 +10,7 @@
 #include "StandardIncludes.hpp"
 #include "WindowController.hpp"
 #include "VK_Pipeline.hpp"
+#include "VK_Types.hpp"
 
 const int MAX_FRAMES_IN_FLIGHT = 2;
 
@@ -29,44 +30,6 @@ const std::vector<const char *> deviceExtensions = {
 
 class Image;
 
-struct Vertex
-{
-    glm::vec2 pos;
-    glm::vec3 color;
-    glm::vec2 texCoord;
-
-    static VkVertexInputBindingDescription getBindingDescription()
-    {
-        VkVertexInputBindingDescription bindingDescription{};
-        bindingDescription.binding = 0;
-        bindingDescription.stride = sizeof(Vertex);
-        bindingDescription.inputRate = VK_VERTEX_INPUT_RATE_VERTEX;
-
-        return bindingDescription;
-    }
-
-    static std::array<VkVertexInputAttributeDescription, 3> getAttributeDescriptions()
-    {
-        std::array<VkVertexInputAttributeDescription, 3> attributeDescriptions{};
-
-        attributeDescriptions[0].binding = 0;
-        attributeDescriptions[0].location = 0;
-        attributeDescriptions[0].format = VK_FORMAT_R32G32_SFLOAT;
-        attributeDescriptions[0].offset = offsetof(Vertex, pos);
-
-        attributeDescriptions[1].binding = 0;
-        attributeDescriptions[1].location = 1;
-        attributeDescriptions[1].format = VK_FORMAT_R32G32B32_SFLOAT;
-        attributeDescriptions[1].offset = offsetof(Vertex, color);
-
-        attributeDescriptions[2].binding = 0;
-        attributeDescriptions[2].location = 2;
-        attributeDescriptions[2].format = VK_FORMAT_R32G32_SFLOAT;
-        attributeDescriptions[2].offset = offsetof(Vertex, texCoord);
-
-        return attributeDescriptions;
-    }
-};
 //
 //
 // struct DrawData
@@ -87,66 +50,6 @@ struct Vertex
 class VulkanBackend : public Singleton<VulkanBackend>
 {
 public:
-    struct Quanta_ImplVulkanH_RenderContext
-    {
-        int Width;
-        int Height;
-        VkSwapchainKHR Swapchain;
-        VkSurfaceKHR Surface;
-        VkSurfaceFormatKHR SurfaceFormat;
-        // VkPresentModeKHR    PresentMode;
-        VkRenderPass RenderPass;
-        VkDescriptorSetLayout descriptorSetLayout;
-        VkPipeline GraphicsPipeline; // The window pipeline may uses a different VkRenderPass than the one passed in ImGui_ImplVulkan_InitInfo
-        VkPipelineLayout PipelineLayout;
-        // bool                UseDynamicRendering;
-        // bool                ClearEnable;
-        // VkClearValue        ClearValue;
-        // uint32_t            FrameIndex;             // Current frame being rendered to (0 <= FrameIndex < FrameInFlightCount)
-        uint32_t ImageIndex; // Number of simultaneous in-flight frames (returned by vkGetSwapchainImagesKHR, usually derived from min_image_count)
-        // uint32_t            SemaphoreCount;         // Number of simultaneous in-flight frames + 1, to be able to use it in vkAcquireNextImageKHR
-        // uint32_t            SemaphoreIndex;         // Current set of swapchain wait semaphores we're using (needs to be distinct from per frame data)
-        // ImGui_ImplVulkanH_Frame* Frames;
-        // ImGui_ImplVulkanH_FrameSemaphores* FrameSemaphores;
-
-        std::vector<VkSemaphore> imageAvailableSemaphores;
-        std::vector<VkSemaphore> renderFinishedSemaphores;
-        std::vector<VkFence> inFlightFences;
-        uint32_t currentFrame = 0;
-
-        Quanta_ImplVulkanH_RenderContext()
-        {
-            memset((void *)this, 0, sizeof(*this));
-            // PresentMode = (VkPresentModeKHR)~0;     // Ensure we get an error if user doesn't set this.
-            // ClearEnable = true;
-        }
-    };
-
-    struct QueueFamilyIndices
-    {
-        std::optional<uint32_t> graphicsFamily;
-        std::optional<uint32_t> presentFamily;
-
-        bool isComplete()
-        {
-            return graphicsFamily.has_value() && presentFamily.has_value();
-        }
-    };
-
-    struct SwapChainSupportDetails
-    {
-        VkSurfaceCapabilitiesKHR capabilities;
-        std::vector<VkSurfaceFormatKHR> formats;
-        std::vector<VkPresentModeKHR> presentModes;
-    };
-
-    struct UniformBufferObject
-    {
-        alignas(16) glm::mat4 model;
-        alignas(16) glm::mat4 view;
-        alignas(16) glm::mat4 proj;
-    };
-
     void SetupVulkan(const char **extensions, uint32_t extensions_count, const std::vector<Vertex> vertices, const std::vector<uint16_t> indices);
 
     void drawFrame(const std::vector<uint16_t> indices);
